added gamma checkbox
