+++ cvat-ui/src/components/annotation-page/canvas/views/canvas2d/image-setups-content.tsx
@@ -20,9 +20,12 @@ import {
     changeContrastLevel,
     changeSaturationLevel,
     changeGridSize,
+    addImageFilter,
+    removeImageFilter,
 } from 'actions/settings-actions';
 import { clamp } from 'utils/math';
 import { GridColor, CombinedState, PlayerSettingsState } from 'reducers';
+import GammaCorrectionImplementation from 'utils/fabric-wrapper/gamma-correciton';
 
 const minGridSize = 5;
 const maxGridSize = 1000;
@@ -181,6 +184,31 @@ export default function ImageSetupsContent(): JSX.Element {
                             >
                                 Reset color settings
                             </Button>
+                        </Col>
+                    </Row>
+                </Col>
+            </Row>
+            <Text>Color settings</Text>
+            <hr />
+            <Row justify='space-around'>
+                <Col span={24}>
+                    <Row className='cvat-image-setups-brightness'>
+                        <Col span={6}>
+                            <Text className='cvat-text-color'> Gamma </Text>
+                        </Col>
+                        <Col span={12}>
+                            <Checkbox
+                                onChange={(e: CheckboxChangeEvent) => {
+                                    if (e.target.checked) {
+                                        dispatch(addImageFilter({
+                                            modifier: new GammaCorrectionImplementation(),
+                                            alias: 'fabric.gamma',
+                                        }));
+                                    } else {
+                                        dispatch(removeImageFilter('fabric.gamma'));
+                                    }
+                                }}
+                            />
                         </Col>
                     </Row>
                 </Col>
